web/requirements/local.txt
-r common.txt
django-nose>=1.4
coverage>=3.7.1
django-debug-toolbar